how-to: source-code: import sending-patches (#73)
* how-to: source-code: import sending-patches

It is common to explain to partners how they can send patches on
our mailing-list. While this guide was previously internal, shipping it
into the public documentation makes more sense as it can help anyone who
wants to contribute to the Ubuntu kernel development.

A separate commit in the private documentation will be created to
redirect to this new page once accepted and live.

Signed-off-by: Agathe Porte <agathe.porte@canonical.com>